requirements.txt
netfilterqueue
scapy
flask
requests
sqlalchemy
python-dotenv
werkzeug
